A simple features data frame with regular hexagons classified into 6 spatial patterns. In order to obtain this classification, we employed a Principal Component Analysis (PCA) to reduce the dimension of a dataset. Then, using only the first thirty-five principal components, we carried out an unsupervised classification to identify groups that represent the distinct spatial patterns within the GGH. Each final cluster can be interpreted as a distinct spatial pattern with unique levels of built-up area density, land use, and street layout. The polygons were classified into: 1 - "High Building Density"; 2 - "Tree-lined Neighborhoods"; 3 - "Medium Building Density"; #' 4 - "Mixed Rural Use"; 5 - "Waterfront Housing"; 6 - "Farms and Rural Neighborhoods"

A simple features data frame with regular hexagons classified into 6 spatial patterns. In order to obtain this classification, we employed a Principal Component Analysis (PCA) to reduce the dimension of a dataset. Then, using only the first thirty-five principal components, we carried out an unsupervised classification to identify groups that represent the distinct spatial patterns within the GGH. Each final cluster can be interpreted as a distinct spatial pattern with unique levels of built-up area density, land use, and street layout. The polygons were classified into: 1 - "High Building Density"; 2 - "Tree-lined Neighborhoods"; 3 - "Medium Building Density"; #' 4 - "Mixed Rural Use"; 5 - "Waterfront Housing"; 6 - "Farms and Rural Neighborhoods"

A simple features data frame with all attributes used during the classification process. In order to obtain this dataset, we extracted features from satellite images and from a road network shapefile. We employed a Principal Component Analysis (PCA) to reduce the dimension of a dataset. Then, using only the first thirty-five principal components, we carried out an unsupervised classification to identify groups that represent the distinct spatial patterns within the GGH.

A simple features data frame with Spatial Access Measures (SAM). In order to obtain this dataset, we transferred the SAM data from the dissemination blocks to the cell grid using a weighted area interpolation. This technique enables us to disaggregate accessibility values to another spatial unit, smoothing variations between polygons. It achieves this by using the area of overlapping geometries to apportion variables.

A simple features data frame with Spatial Access Measures (SAM) and the spatial patterns classification.

Usage

data(grid_classified)

data(grid_classified)

data(grid_classified)

data(grid_classified)

Format

A simple features data frame with 53,609 rows and 2 variables:

id

unique identifier for each polygon.

Cluster

space pattern classification

A simple features data frame with 53,609 rows and 228 variables:

id

unique identifier for each polygon.

BP_min

minimum value of built-up area

BP_max

maximum value of built-up area

BP_mean

average value of built-up area

BP_sum

sum of built-up area

BP_std

standard deviation of built-up area

BP_median

median value of built-up area

BP_majorit

majority value of built-up area

HE_min

minimum value of building height

HE_max

maximum value of building height

HE_mean

average value of building height

HE_sum

sum of building height

HE_std

standard deviation of building height

HE_median

median value of building height

HE_majorit

majority value of building height

VO_min

minimum value of buildings volume

VO_max

maximum value of buildings volume

VO_mean

average value of buildings volume

VO_sum

sum of buildings volume

VO_std

standard deviation of buildings volume

VO_median

median value of buildings volume

VO_majorit

majority value of buildings volume

NRC

Percentage of non-residential and non-commercial area

RES

Percentage of residential area

COM

Percentage of commercial area

TR_min

minimum value of tree fraction

TR_max

maximum value of tree fraction

TR_mean

average value of tree fraction

TR_sum

sum of tree fraction

TR_std

standard deviation of tree fraction

TR_median

median value of tree fraction

TR_majorit

majority value of tree fraction

CR_min

minimum value of crop fraction

CR_max

maximum value of crop fraction

CR_mean

average value of crop fraction

CR_sum

sum of crop fraction

CR_std

standard deviation of crop fraction

CR_median

median value of crop fraction

CR_majorit

majority value of crop fraction

GR_min

minimum value of grass fraction

GR_max

maximum value of grass fraction

GR_mean

average value of grass fraction

GR_sum

sum of grass fraction

GR_std

standard deviation of grass fraction

GR_median

median value of grass fraction

GR_majorit

majority value of grass fraction

WA_min

minimum value of water fraction

WA_max

maximum value of water fraction

WA_mean

average value of water fraction

WA_sum

sum of water fraction

WA_std

standard deviation of water fraction

WA_median

median value of water fraction

WA_majorit

majority value of water fraction

SH_min

minimum value of shrub area

SH_max

maximum value of shrub area

SH_mean

average value of shrub area

SH_sum

sum of shrub area

SH_std

standard deviation of shrub area

SH_median

median value of shrub area

SH_majorit

majority value of shrub area

SI_mean

average value of road sinuosity

SI_sum

sum of road sinuosity

SI_min

minimum value of road sinuosity

SI_max

maximum value of road sinuosity

SI_std

standard deviation of road sinuosity

LE_mean

average value of road length

LE_sum

sum of road length

LE_min

minimum value of road length

LE_max

maximum value of road length

LE_std

standard deviation of road length

NRo

road count

NIn

number of road intersections

DInBP

road intersections divided by the sum of built-up area (NRo/BP_sum)

DInRo

road intersections divided by count of roads (NRo/NRi)

NM_BP_min

average of the neighbors for the variable BP_min

NM_BP_max

average of the neighbors for the variable BP_max

NM_BP_mean

average of the neighbors for the variable BP_mean

NM_BP_sum

average of the neighbors for the variable BP_sum

NM_BP_std

average of the neighbors for the variable BP_std

NM_BP_medi

average of the neighbors for the variable BP_medi

NM_BP_majo

average of the neighbors for the variable BP_majo

NM_HE_min

average of the neighbors for the variable HE_min

NM_HE_max

average of the neighbors for the variable HE_max

NM_HE_mean

average of the neighbors for the variable HE_mean

NM_HE_sum

average of the neighbors for the variable HE_sum

NM_HE_std

average of the neighbors for the variable HE_std

NM_HE_medi

average of the neighbors for the variable HE_medi

NM_HE_majo

average of the neighbors for the variable HE_majo

NM_VO_min

average of the neighbors for the variable VO_min

NM_VO_max

average of the neighbors for the variable VO_max

NM_VO_mean

average of the neighbors for the variable VO_mean

NM_VO_sum

average of the neighbors for the variable VO_sum

NM_VO_std

average of the neighbors for the variable VO_std

NM_VO_medi

average of the neighbors for the variable VO_medi

NM_VO_majo

average of the neighbors for the variable VO_majo

NM_NRC

average of the neighbors for the variable Percentage of non-residential and non-commercial area

NM_RES

average of the neighbors for the variable Percentage of residential area

NM_COM

average of the neighbors for the variable Percentage of commercial area

NM_TR_min

average of the neighbors for the variable TR_min

NM_TR_max

average of the neighbors for the variable TR_max

NM_TR_mean

average of the neighbors for the variable TR_mean

NM_TR_sum

average of the neighbors for the variable TR_sum

NM_TR_std

average of the neighbors for the variable TR_std

NM_TR_medi

average of the neighbors for the variable TR_medi

NM_TR_majo

average of the neighbors for the variable TR_majo

NM_CR_min

average of the neighbors for the variable CR_min

NM_CR_max

average of the neighbors for the variable CR_max

NM_CR_mean

average of the neighbors for the variable CR_mean

NM_CR_sum

average of the neighbors for the variable CR_sum

NM_CR_std

average of the neighbors for the variable CR_std

NM_CR_medi

average of the neighbors for the variable CR_medi

NM_CR_majo

average of the neighbors for the variable CR_majo

NM_GR_min

average of the neighbors for the variable GR_min

NM_GR_max

average of the neighbors for the variable GR_max

NM_GR_mean

average of the neighbors for the variable GR_mean

NM_GR_sum

average of the neighbors for the variable GR_sum

NM_GR_std

average of the neighbors for the variable GR_std

NM_GR_medi

average of the neighbors for the variable GR_medi

NM_GR_majo

average of the neighbors for the variable GR_majo

NM_WA_min

average of the neighbors for the variable WA_min

NM_WA_max

average of the neighbors for the variable WA_max

NM_WA_mean

average of the neighbors for the variable WA_mean

NM_WA_sum

average of the neighbors for the variable WA_sum

NM_WA_std

average of the neighbors for the variable WA_std

NM_WA_medi

average of the neighbors for the variable WA_medi

NM_WA_majo

average of the neighbors for the variable WA_majo

NM_SH_min

average of the neighbors for the variable SH_min

NM_SH_max

average of the neighbors for the variable SH_max

NM_SH_mean

average of the neighbors for the variable SH_mean

NM_SH_sum

average of the neighbors for the variable SH_sum

NM_SH_std

average of the neighbors for the variable SH_std

NM_SH_medi

average of the neighbors for the variable SH_medi

NM_SH_majo

average of the neighbors for the variable SH_majo

DNM_BP_min

normalized difference between the cell and its neighbors for the variable BP_min

DNM_BP_max

normalized difference between the cell and its neighbors for the variable BP_max

DNM_BP_mea

normalized difference between the cell and its neighbors for the variable BP_mea

DNM_BP_sum

normalized difference between the cell and its neighbors for the variable BP_sum

DNM_BP_std

normalized difference between the cell and its neighbors for the variable BP_std

DNM_BP_med

normalized difference between the cell and its neighbors for the variable BP_med

DNM_BP_maj

normalized difference between the cell and its neighbors for the variable BP_maj

DNM_HE_min

normalized difference between the cell and its neighbors for the variable HE_min

DNM_HE_max

normalized difference between the cell and its neighbors for the variable HE_max

DNM_HE_mea

normalized difference between the cell and its neighbors for the variable HE_mea

DNM_HE_sum

normalized difference between the cell and its neighbors for the variable HE_sum

DNM_HE_std

normalized difference between the cell and its neighbors for the variable HE_std

DNM_HE_med

normalized difference between the cell and its neighbors for the variable HE_med

DNM_HE_maj

normalized difference between the cell and its neighbors for the variable HE_maj

DNM_VO_min

normalized difference between the cell and its neighbors for the variable VO_min

DNM_VO_max

normalized difference between the cell and its neighbors for the variable VO_max

DNM_VO_mea

normalized difference between the cell and its neighbors for the variable VO_mea

DNM_VO_sum

normalized difference between the cell and its neighbors for the variable VO_sum

DNM_VO_std

normalized difference between the cell and its neighbors for the variable VO_std

DNM_VO_med

normalized difference between the cell and its neighbors for the variable VO_med

DNM_VO_maj

normalized difference between the cell and its neighbors for the variable VO_maj

DNM_NRC

normalized difference between the cell and its neighbors for the variable Percentage of non-residential and non-commercial area

DNM_RES

normalized difference between the cell and its neighbors for the variable Percentage of residential area

DNM_COM

normalized difference between the cell and its neighbors for the variable Percentage of commercial area

DNM_TR_min

normalized difference between the cell and its neighbors for the variable TR_min

DNM_TR_max

normalized difference between the cell and its neighbors for the variable TR_max

DNM_TR_mea

normalized difference between the cell and its neighbors for the variable TR_mea

DNM_TR_sum

normalized difference between the cell and its neighbors for the variable TR_sum

DNM_TR_std

normalized difference between the cell and its neighbors for the variable TR_std

DNM_TR_med

normalized difference between the cell and its neighbors for the variable TR_med

DNM_TR_maj

normalized difference between the cell and its neighbors for the variable TR_maj

DNM_CR_min

normalized difference between the cell and its neighbors for the variable CR_min

DNM_CR_max

normalized difference between the cell and its neighbors for the variable CR_max

DNM_CR_mea

normalized difference between the cell and its neighbors for the variable CR_mea

DNM_CR_sum

normalized difference between the cell and its neighbors for the variable CR_sum

DNM_CR_std

normalized difference between the cell and its neighbors for the variable CR_std

DNM_CR_med

normalized difference between the cell and its neighbors for the variable CR_med

DNM_CR_maj

normalized difference between the cell and its neighbors for the variable CR_maj

DNM_GR_min

normalized difference between the cell and its neighbors for the variable GR_min

DNM_GR_max

normalized difference between the cell and its neighbors for the variable GR_max

DNM_GR_mea

normalized difference between the cell and its neighbors for the variable GR_mea

DNM_GR_sum

normalized difference between the cell and its neighbors for the variable GR_sum

DNM_GR_std

normalized difference between the cell and its neighbors for the variable GR_std

DNM_GR_med

normalized difference between the cell and its neighbors for the variable GR_med

DNM_GR_maj

normalized difference between the cell and its neighbors for the variable GR_maj

DNM_WA_min

normalized difference between the cell and its neighbors for the variable WA_min

DNM_WA_max

normalized difference between the cell and its neighbors for the variable WA_max

DNM_WA_mea

normalized difference between the cell and its neighbors for the variable WA_mea

DNM_WA_sum

normalized difference between the cell and its neighbors for the variable WA_sum

DNM_WA_std

normalized difference between the cell and its neighbors for the variable WA_std

DNM_WA_med

normalized difference between the cell and its neighbors for the variable WA_med

DNM_WA_maj

normalized difference between the cell and its neighbors for the variable WA_maj

DNM_SH_min

normalized difference between the cell and its neighbors for the variable SH_min

DNM_SH_max

normalized difference between the cell and its neighbors for the variable SH_max

DNM_SH_mea

normalized difference between the cell and its neighbors for the variable SH_mea

DNM_SH_sum

normalized difference between the cell and its neighbors for the variable SH_sum

DNM_SH_std

normalized difference between the cell and its neighbors for the variable SH_std

DNM_SH_med

normalized difference between the cell and its neighbors for the variable SH_med

DNM_SH_maj

normalized difference between the cell and its neighbors for the variable SH_maj

PC1

Principal Component 1

PC2

Principal Component 2

PC3

Principal Component 3

PC4

Principal Component 4

PC5

Principal Component 5

PC6

Principal Component 6

PC7

Principal Component 7

PC8

Principal Component 8

PC9

Principal Component 9

PC10

Principal Component 10

PC11

Principal Component 11

PC12

Principal Component 12

PC13

Principal Component 13

PC14

Principal Component 14

PC15

Principal Component 15

PC16

Principal Component 16

PC17

Principal Component 17

PC18

Principal Component 18

PC19

Principal Component 19

PC20

Principal Component 20

PC21

Principal Component 21

PC22

Principal Component 22

PC23

Principal Component 23

PC24

Principal Component 24

PC25

Principal Component 25

PC26

Principal Component 26

PC27

Principal Component 27

PC28

Principal Component 28

PC29

Principal Component 29

PC30

Principal Component 30

PC31

Principal Component 31

PC32

Principal Component 32

PC33

Principal Component 33

PC34

Principal Component 34

PC35

Principal Component 35

Cluster

space pattern classification

A simple features data frame with 53,609 rows and 5 variables:

id

unique identifier for each polygon.

cy_ac_emp

normalised value of a cell's access to employment (Cycling mode)

pto_ac_emp

normalised value of a cell's access to employment (Publit transit (off peak) mode)

ptp_ac_emp

normalised value of a cell's access to employment (Publit transit (on peak) mode)

wa_ac_emp

normalised value of a cell's access to employment (Walking mode)

A simple features data frame with 13,097 rows and 5 variables:

id

unique identifier for each polygon.

Cluster

space pattern classification

cy_ac_emp

normalised value of a cell's access to employment (Cycling mode)

pto_ac_emp

normalised value of a cell's access to employment (Publit transit (off peak) mode)

ptp_ac_emp

normalised value of a cell's access to employment (Publit transit (on peak) mode)

wa_ac_emp

normalised value of a cell's access to employment (Walking mode)

Source

Global Human Settlement Layer (GHSL) (https://ghsl.jrc.ec.europa.eu/dataToolsOverview.php#inline-nav-R2023) Accessed: December 14, 2023

OpenStreetMap (https://download.geofabrik.de/) Accessed: December 14, 2023

Copernicus Global Land Service (CGLS) (https://land.copernicus.eu/global/index.html) Accessed: December 14, 2023

Global Human Settlement Layer (GHSL) (https://ghsl.jrc.ec.europa.eu/dataToolsOverview.php#inline-nav-R2023) Accessed: December 14, 2023

OpenStreetMap (https://download.geofabrik.de/) Accessed: December 14, 2023

Copernicus Global Land Service (CGLS) (https://land.copernicus.eu/global/index.html) Accessed: December 14, 2023

Statistics Canada (https://www150.statcan.gc.ca/n1/en/catalogue/27260001) Accessed: December 14, 2023

Statistics Canada (https://www150.statcan.gc.ca/n1/en/catalogue/27260001) Accessed: December 14, 2023

Examples

 data(grid_classified)
 clusters <- grid_classified$Cluster
 data(grid_with_attributes)
 built_up_mean <- grid_with_attributes$BP_mean
 data(grid_sam)
 cycling <- grid_sam$cy_ac_emp
 data(grid_sam_classified)
 cycling <- grid_sam_classified$cy_ac_emp